Address feedback * Update some padding and styling of TCs * Clean up imports and unused methods * Rename unclear classes * Change TC's public icon to another icon for clarity
