+++ src/fontawesome.tsx
@@ -8,12 +8,11 @@ import {
     faLock,
     faMapMarkerAlt,
     faPlusCircle,
-    faUnlock,
     faEdit,
     faCopy,
     faTrash,
     faEllipsisH,
-    faSearch,
+    faSearch, faLockOpen
 } from "@fortawesome/free-solid-svg-icons";
 import {faQuestionCircle, faStar} from "@fortawesome/free-regular-svg-icons";
 import {faGithub} from "@fortawesome/free-brands-svg-icons";
@@ -25,7 +24,7 @@ library.add(
     faChevronUp,
     faChevronDown,
     faLock,
-    faUnlock,
+    faLockOpen,
     faBorderAll,
     faEdit,
     faCopy,
