+++ src/stylesheets/footer.scss
@@ -33,7 +33,7 @@
     }
   }
 
-  .sns {
+  .links {
     float: right;
     width: 50%;
     display: flex;
