+++ src/stylesheets/trecipeCard.scss
@@ -4,6 +4,7 @@
   font-family: $default-font-family;
   background-color: $tc-white;
   border:  1px solid $light-gray;
+  border-radius: 8px;
   width: 100%;
   height: 300px;
   display: flex;
@@ -15,6 +16,7 @@
   background-position: center;
   background-repeat: no-repeat;
   height: 67%;
+  border-radius: 8px 8px 0 0;
 
   .tcHeader {
     background-image: linear-gradient(180deg,
@@ -82,6 +84,7 @@
   height: 10px;
   position: relative;
   background-color: $tc-white;
+  border-radius: 0 0 8px 8px;
 
   .tcFiller {
     background-color: $powder-blue;
