+++ src/stylesheets/variables.scss
@@ -27,4 +27,4 @@ $default-font-size: 1rem;
 $title-font-size: 2rem;
 
 // --- Element Widths ---//
-$tc-max-width: 350px;
+$tc-min-width: 350px;
